- [Vulkan] Fixed a swapchain bug where the cmd index would not increment correctly (increased performance) - [Vulkan] Optimised descriptor set capacity reseting - [DirectXShaderCompiler] Updated to latest and added to the libraries used list (in the about window) - [Misc] Added optimise on/off macros

+++ Runtime/RHI/Vulkan/Vulkan_PipelineState.cpp
@@ -118,7 +118,7 @@ namespace Spartan
                         if (!texture)
                             continue;
 
-                        VkImageLayout layout = vulkan_image_layout[texture->GetLayout()];
+                        VkImageLayout layout = vulkan_image_layout[static_cast<uint8_t>(texture->GetLayout())];
 
                         VkAttachmentDescription attachment_desc  = {};
                         attachment_desc.format                   = vulkan_format[texture->GetFormat()];
@@ -141,7 +141,7 @@ namespace Spartan
             // Depth
             if (render_target_depth_texture)
             {
-                VkImageLayout layout = vulkan_image_layout[render_target_depth_texture->GetLayout()];
+                VkImageLayout layout = vulkan_image_layout[static_cast<uint8_t>(render_target_depth_texture->GetLayout())];
 
                 VkAttachmentDescription attachment_desc  = {};
                 attachment_desc.format                   = vulkan_format[render_target_depth_texture->GetFormat()];
